Notifications plugin should not do anything if it is running within a parallel child (closes #776)
